Updated CHANGELOG.md Fixed test coverage CI Moved `trace_err` to services
