+++ crates/services/src/lib.rs
@@ -40,6 +40,24 @@ pub mod stream {
     impl<S> IntoBoxStream for S where S: Stream + Send + Sync + 'static {}
 }
 
+/// Helper trait to trace errors
+pub trait TraceErr {
+    /// Trace an error with a message.
+    fn trace_err(self, msg: &str) -> Self;
+}
+
+impl<T, E> TraceErr for Result<T, E>
+where
+    E: Display,
+{
+    fn trace_err(self, msg: &str) -> Self {
+        if let Err(e) = &self {
+            tracing::error!("{} {}", msg, e);
+        }
+        self
+    }
+}
+
 pub use service::{
     EmptyShared,
     RunnableService,
@@ -51,6 +69,7 @@ pub use state::{
     State,
     StateWatcher,
 };
+use std::fmt::Display;
 pub use sync::{
     Shared,
     SharedMutex,
