+++ crates/trace/src/lib.rs
@@ -2,32 +2,11 @@
 #![deny(clippy::cast_possible_truncation)]
 
 use ctor::ctor;
-use std::{
-    env::var,
-    fmt::Display,
-};
+use std::env::var;
 use tracing_subscriber::{
     fmt::format,
     EnvFilter,
 };
-
-pub use tracing;
-
-pub trait TraceErr {
-    fn trace_err(self, msg: &str) -> Self;
-}
-
-impl<T, E> TraceErr for Result<T, E>
-where
-    E: Display,
-{
-    fn trace_err(self, msg: &str) -> Self {
-        if let Err(e) = &self {
-            tracing::error!("{} {}", msg, e);
-        }
-        self
-    }
-}
 
 #[ctor]
 pub static TRACE: () = {
